Partial solution for issue 34.
This maintains undo buffers for documents when switching back and
forth, but the buffer is still cleared out when hitting the "Test"
button. Fixing that is more tricky.

This also adds support for font size changing (as in issue 22)
with the default (non-pygments-based) highlighter. The main
trouble still remaining here is that resizing the fonts does
cause undo/redo issues. So pick a font size and stick with it.